Add files via upload
PowerShell is a powerful scripting language that can be used to configure and manage firewalls on Windows machines. It allows you to automate firewall configuration tasks, such as adding or removing firewall rules, enabling or disabling the firewall, and configuring advanced firewall settings.

There are several cmdlets in PowerShell that are specifically designed for firewall configuration. Some of the most commonly used cmdlets include:

New-NetFirewallRule: This cmdlet allows you to create a new firewall rule.
Set-NetFirewallRule: This cmdlet allows you to modify an existing firewall rule.
Get-NetFirewallRule: This cmdlet allows you to retrieve information about a firewall rule.
Enable-NetFirewallRule: This cmdlet allows you to enable a firewall rule.
Disable-NetFirewallRule: This cmdlet allows you to disable a firewall rule.
Set-NetFirewallProfile: This cmdlet allows you to configure the firewall profile settings.