Add files via upload
This script prompts the user for information about the new user to be created, creates a new UserPrincipal object, sets the necessary properties for the user, saves the user object in Active Directory, and then moves the user object to the designated Organizational Unit (OU). Finally, a confirmation message is displayed. This script can be easily modified and adapted to automate other tasks in a Windows environment.