Add files via upload
This script will search for available updates that have not been installed, download and install them, and then reboot the system if required. The script uses the Microsoft.Update.ServiceManager COM object to manage Windows updates and the UpdateInstaller and UpdateDownloader COM objects to install and download updates respectively. The script also checks if any of the updates require a reboot and restarts the system if necessary.